CDROM: Ignore sectors with channel number 255
This was inspired by the fix in Duckstation :
stenzek/duckstation@0710e3b

Some games have junk audio sectors with a channel number of 255.
If these are not skipped, then they will play wrong file.

This was tested on "Blue's Clues : Blue's Big Musical" and
it fixed the missing audio there.

Taxi 2 is also said to be affected by this.